Expose a camera context for the 3d scenes.
This context links the camera of the different 3d scenes.
It's implemented by using a reducer and a shared state.
If a camera moves, an action is dispatched, and all the scenes will
react by updating their camera. We pass the ID of the scenes whose
camera moved to avoid re-rendering this scene.